modify

+++ lightning_attn_cute/triton_compare/triton_kernels.py
@@ -54,7 +54,7 @@ def fwd_kernel_v4(
         v = tl.load(V_start + vo_off, mask=block_off[:, None] < n, other=0.0).to(tl.float16)
 
 
-        qk = tl.dot(q, k_t)
+        qk = tl.dot(q, k_t).to(tl.float16)
         o_intra = tl.dot(qk * diag_decay, v)
         o_inter = tl.dot(q, kv) * q_decay
         o = o_intra + o_inter
